pyproject.toml
[project]
name = "python-project-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery[redis]>=5.6.3",
    "django>=6.0.4",
    "django-stubs[compatible-mypy]>=6.0.3",
    "fastapi>=0.136.1",
    "jinja2>=3.1.6",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pytest-coverage>=0.0",
    "pytest-env>=1.6.0",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "typer>=0.25.0",
    "uvicorn>=0.46.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.3.0",
    "ipykernel>=6.30.1",
    "mypy>=1.20.2",
    "pytest>=9.0.3",
]

[tool.uv]
package = true

[project.scripts]
python-project-starter = "project:cli"

